Extract marketing pack purchase logic in hook
