+++ newIDE/app/src/AssetStore/CreditsPackages/CreditsPackageStoreContext.js
@@ -122,6 +122,7 @@ export const CreditsPackageStoreStateProvider = ({
 
   const openCreditsPackageDialog = React.useCallback(
     (options: ?CreditsPackageDialogOpeningOptions) => {
+      console.log('hou');
       if (!creditsPackageListingDatas) return;
 
       setMissingCredits(
@@ -184,7 +185,7 @@ export const CreditsPackageStoreStateProvider = ({
     []
   );
 
-  const CreditsPackageStoreState = React.useMemo(
+  const creditsPackageStoreState = React.useMemo(
     () => ({
       creditsPackageListingDatas,
       fetchCreditsPackages,
@@ -204,7 +205,7 @@ export const CreditsPackageStoreStateProvider = ({
   );
 
   return (
-    <CreditsPackageStoreContext.Provider value={CreditsPackageStoreState}>
+    <CreditsPackageStoreContext.Provider value={creditsPackageStoreState}>
       {children}
       {isCreditsPackageDialogOpen && (
         <CreditsPackagesDialog
